+++ src/scenes/App/NoSelectedProvider/NoSelectedProvider.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavigationCard } from '@entur/layout';
 import { AdjustmentsIcon } from '@entur/icons';
-import { SelectProvider } from './SelectProvider';
+import { SelectProvider } from '../SelectProvider/SelectProvider';
 import { useIntl } from 'react-intl';
 
 export const NoSelectedProvider = () => {
